% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{blr}
\alias{blr}
\title{Logistic regression}
\usage{
blr(
  formula,
  data,
  subset,
  na.action,
  nu = 1,
  mu0 = NULL,
  tau.sq = NULL,
  start = NULL,
  control = NULL,
  contrasts = NULL,
  model = TRUE
)
}
\description{
Logistic regression
}
