% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{brlm}
\alias{brlm}
\title{Robust linear regression}
\usage{
brlm(
  formula,
  data,
  subset,
  na.action,
  nu = 7,
  mu0 = NULL,
  tau.sq = NULL,
  start = NULL,
  control = NULL,
  contrasts = NULL,
  model = TRUE
)
}
\description{
Robust linear regression
}
